refactor(cip30): hoist WalletMethodNames to Wallet
